bitops: Introduce the for_each_set_port_word macro
This macro iterates for each group of bits (port word) with set bits,
within a bitmap memory region. For each iteration, "port_word" is set to
the found port word index, "word_index" is set to the word index of the
bitmap containing the found port word, and "word_offset" is set to the
bit offset of the found port word within the respective bitmap word.

Signed-off-by: William Breathitt Gray <vilhelm.gray@gmail.com>